requirements.txt
flask==2.1.1
requests==2.27.1
Flask-Scss==0.5
Flask-Assets==2.0
livereload==2.6.3
flask-bcrypt==1.0.1
flask_sqlalchemy==2.5.1
flask-wtf==1.0.1
email_validator==1.2.1
pillow==9.1.1
sqlalchemy-utils==0.38.2
flask-socketio==5.2.0
numpy==1.23.1
flask_mail==0.9.1
pandas==1.4.4
python-socketio[client]==5.7.0